# artifacts/api-server/src/routes/closing-checklist.ts
import { Router, type IRouter } from "express";
import { prisma } from "@workspace/db";
import { requireAuth } from "../middlewares/auth.js";
import { userCanAccessProject, userCanWriteProject } from "../lib/projectAccess.js";
import { recordAudit } from "../lib/audit.js";

const router: IRouter = Router();
router.use(requireAuth);

export const CLOSING_CHECKLIST_DEFAULTS: Array<{ key: string; label: string; sortOrder: number }> = [
  { key: "BAST_SIGNED", label: "BAST signed by client", sortOrder: 10 },
  { key: "FINAL_REPORT_DELIVERED", label: "Final report delivered to client", sortOrder: 20 },
  { key: "INVOICE_ISSUED", label: "Invoice issued", sortOrder: 30 },
  { key: "PAYMENT_RECEIVED", label: "Payment received in full", sortOrder: 40 },
  { key: "ALL_TIMESHEETS_APPROVED", label: "All timesheets approved", sortOrder: 50 },
  { key: "ALL_EXPENSES_FINALIZED", label: "All expenses approved/rejected (none pending)", sortOrder: 60 },
  { key: "SURVEY_SENT", label: "Client satisfaction survey sent", sortOrder: 70 },
  { key: "LESSONS_LEARNED", label: "Lessons learned documented", sortOrder: 80 },
];

async function ensureChecklist(projectId: string) {
  const existing = await prisma.projectClosingChecklistItem.findMany({ where: { projectId } });
  if (existing.length >= CLOSING_CHECKLIST_DEFAULTS.length) return existing;
  const existingKeys = new Set(existing.map((e) => e.key));
  for (const d of CLOSING_CHECKLIST_DEFAULTS) {
    if (!existingKeys.has(d.key)) {
      await prisma.projectClosingChecklistItem.create({
        data: {
          projectId,
          key: d.key,
          label: d.label,
          sortOrder: d.sortOrder,
          status: "PENDING",
        },
      });
    }
  }
  return prisma.projectClosingChecklistItem.findMany({
    where: { projectId },
    orderBy: [{ sortOrder: "asc" }],
  });
}

function serialize(it: Awaited<ReturnType<typeof prisma.projectClosingChecklistItem.findFirstOrThrow>>) {
  return {
    id: it.id,
    projectId: it.projectId,
    key: it.key,
    label: it.label,
    status: it.status,
    note: it.note,
    completedAt: it.completedAt ? it.completedAt.toISOString() : null,
    completedById: it.completedById,
    sortOrder: it.sortOrder,
  };
}

/**
 * Returns an error message when the checklist item's evidence requirement is
 * not met, or null when the item may be marked DONE.
 */
async function checkDoneRequirement(projectId: string, key: string): Promise<string | null> {
  if (key === "BAST_SIGNED") {
    const bast = await prisma.document.findFirst({
      where: { projectId, type: "BAST", isLatest: true },
      select: { id: true },
    });
    if (!bast) {
      return "Upload or link a BAST document first (Documents tab) before marking this item done.";
    }
  }
  if (key === "FINAL_REPORT_DELIVERED") {
    const report = await prisma.document.findFirst({
      where: { projectId, type: "REPORT", isLatest: true },
      select: { id: true },
    });
    if (!report) {
      return "Upload or link the final report (document type REPORT) before marking this item done.";
    }
  }
  if (key === "INVOICE_ISSUED") {
    const invoiceDoc = await prisma.document.findFirst({
      where: { projectId, type: "INVOICE", isLatest: true },
      select: { id: true },
    });
    if (!invoiceDoc) {
      const invoicedMilestone = await prisma.billingMilestone.findFirst({
        where: { projectId, status: { in: ["INVOICED", "PAID"] } },
        select: { id: true },
      });
      if (!invoicedMilestone) {
        return "Attach an invoice document or mark a billing milestone as invoiced before marking this item done.";
      }
    }
  }
  return null;
}

router.get("/projects/:id/closing-checklist", async (req, res) => {
  const projectId = String(req.params.id);
  if (!(await userCanAccessProject(projectId, req.user!))) {
    res.status(404).json({ error: "Not found" });
    return;
  }
  const items = await ensureChecklist(projectId);
  res.json(items.sort((a, b) => a.sortOrder - b.sortOrder).map(serialize));
});

router.patch("/projects/:id/closing-checklist/:itemId", async (req, res) => {
  const projectId = String(req.params.id);
  const itemId = String(req.params.itemId);
  if (!(await userCanWriteProject(projectId, req.user!))) {
    res.status(403).json({ error: "Forbidden" });
    return;
  }
  const item = await prisma.projectClosingChecklistItem.findUnique({ where: { id: itemId } });
  if (!item || item.projectId !== projectId) {
    res.status(404).json({ error: "Not found" });
    return;
  }
  const { status, note } = req.body || {};
  const data: Record<string, unknown> = {};
  if (status !== undefined) {
    if (!["PENDING", "DONE", "NA"].includes(String(status))) {
      res.status(400).json({ error: "status must be PENDING|DONE|NA" });
      return;
    }
    // Evidence-based items cannot be marked DONE unless the underlying
    // document/link actually exists. NA stays allowed as the explicit
    // "not applicable" escape hatch (e.g. non-commercial projects).
    if (status === "DONE") {
      const requirement = await checkDoneRequirement(projectId, item.key);
      if (requirement) {
        res.status(400).json({ error: requirement });
        return;
      }
      // F6: lessons learned must actually be written down — marking the item
      // DONE requires a non-empty note (incoming note wins over stored note).
      if (item.key === "LESSONS_LEARNED") {
        const effectiveNote = note !== undefined ? String(note ?? "").trim() : String(item.note ?? "").trim();
        if (!effectiveNote) {
          res.status(400).json({
            error: "Write the lessons learned in the note field before marking this item done.",
          });
          return;
        }
      }
    }
    data.status = String(status);
    if (status === "DONE" || status === "NA") {
      data.completedAt = new Date();
      data.completedById = req.user!.sub;
    } else {
      data.completedAt = null;
      data.completedById = null;
    }
  }
  if (note !== undefined) data.note = note || null;
  const updated = await prisma.projectClosingChecklistItem.update({ where: { id: itemId }, data });
  res.json(serialize(updated));
});

export default router;