# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SecureProfit Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{id}/documents:
    get:
      operationId: listProjectDocuments
      tags: [documents]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: includeHistory, in: query, required: false, schema: { type: boolean } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Document" }
    post:
      operationId: createProjectDocument
      tags: [documents]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateDocumentBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Document" }
  /documents/{id}/versions:
    get:
      operationId: listDocumentVersions
      tags: [documents]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Document" }
  /projects/{id}/closing-checklist:
    get:
      operationId: getProjectClosingChecklist
      tags: [projects]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ClosingChecklistItem" }
  /projects/{id}/closing-checklist/{itemId}:
    patch:
      operationId: updateProjectClosingChecklistItem
      tags: [projects]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: itemId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateClosingChecklistItemBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ClosingChecklistItem" }
...
  /documents/{id}:
    delete:
      operationId: deleteDocument
      tags: [documents]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SuccessMessage" }
...
components:
  schemas:
...
    SuccessMessage:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
        message: { type: string }
...
    DocumentType:
      type: string
      enum: [BAST, INVOICE, CONTRACT, REPORT, OTHER]
    DocumentKind:
      type: string
      enum: [FILE, LINK]
...
    Document:
      type: object
      required: [id, projectId, type, fileName, fileUrl, uploadedAt, version, isLatest]
      properties:
        id: { type: string }
        projectId: { type: string }
        type: { $ref: "#/components/schemas/DocumentType" }
        kind: { $ref: "#/components/schemas/DocumentKind" }
        fileName: { type: string }
        fileUrl: { type: string }
        invoiceNumber: { type: string, nullable: true }
        invoiceAmount: { type: number, nullable: true }
        invoiceStatus: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        billingMilestoneId: { type: string, nullable: true }
        billingMilestoneName: { type: string, nullable: true }
        uploadedById: { type: string, nullable: true }
        uploadedByName: { type: string, nullable: true }
        uploadedAt: { type: string }
        version: { type: integer }
        parentDocumentId: { type: string, nullable: true }
        isLatest: { type: boolean }
    ClosingChecklistItem:
      type: object
      required: [id, projectId, key, label, status, sortOrder]
      properties:
        id: { type: string }
        projectId: { type: string }
        key: { type: string }
        label: { type: string }
        status: { type: string, enum: [PENDING, DONE, NA] }
        note: { type: string, nullable: true }
        completedAt: { type: string, nullable: true }
        completedById: { type: string, nullable: true }
        sortOrder: { type: integer }
...
    UpdateClosingChecklistItemBody:
      type: object
      properties:
        status: { type: string, enum: [PENDING, DONE, NA] }
        note: { type: string, nullable: true }
...
    CreateDocumentBody:
      type: object
      required: [type, fileName, fileUrl]
      properties:
        type: { $ref: "#/components/schemas/DocumentType" }
        kind: { $ref: "#/components/schemas/DocumentKind" }
        fileName: { type: string }
        fileUrl: { type: string }
        invoiceNumber: { type: string }
        invoiceAmount: { type: number }
        invoiceStatus: { type: string }
        notes: { type: string }
        billingMilestoneId: { type: string, nullable: true }
...