# lib/api-zod/src/generated/types/documentKind.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SecureProfit Hub API
 * OpenAPI spec version: 0.1.0
 */

export type DocumentKind = (typeof DocumentKind)[keyof typeof DocumentKind];

export const DocumentKind = {
  FILE: "FILE",
  LINK: "LINK",
} as const;